Moving most of the Map logic into a Map component
